9559: split out new computeSubmitTrialSessionDataAction into new file, consolidated and exported associated helper function to be accessible by both
